Moves input-related code to its own module with one small change
Input impl now provides a `try_recv()` method, which simply delegates to
the underlying mpsc recv channel. I view this as preferable to exposing
raw access to the actual channel in the main loop.
